inference/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
opencv-python-headless==4.8.1.78
onnxruntime==1.16.1
numpy==1.24.3
python-multipart==0.0.6

flask==2.3.3
requests==2.31.0
pillow==10.0.0
ultralytics>=8.0.0

websockets==12.0